Moved API Models in the Common project
